Split jsonrpc2.go file into multiple files
